bug fix
the cpmpiler in linux works diffrently, so I had to change the struct
and enums with some sort of 'defif'

also I had a problem with input of negative number to take from the
heaps, so I added an if for that.

also I checked the out files and noticed that we needed to add a ' '
(space) where the heap was empty. so I did.